refactor: split classtable and settings sections
