+++ lib/page/setting/sections/cancel_button.dart
@@ -1,0 +1,19 @@
+// Copyright 2023-2025 BenderBlog Rodriguez and contributors
+// Copyright 2025 Traintime PDA authors.
+// SPDX-License-Identifier: MPL-2.0
+
+part of '../setting_sections.dart';
+
+class _CancelButton extends StatelessWidget {
+  @override
+  Widget build(BuildContext context) {
+    return TextButton(
+      style: TextButton.styleFrom(
+        backgroundColor: Theme.of(context).colorScheme.primary,
+        foregroundColor: Theme.of(context).colorScheme.onPrimary,
+      ),
+      onPressed: () => Navigator.pop(context),
+      child: Text(FlutterI18n.translate(context, "cancel")),
+    );
+  }
+}
